sua conflic

+++ src/main/java/org/example/test/service/OrderService.java
@@ -42,7 +42,7 @@ public class OrderService {
         return orderRepository.findAll();
     }
 
-    public Order getOrderById(int id) {
+    public Order getOrderById(long id) {
         Optional<Order> optional = orderRepository.findById(id);
         Order order = null;
         if (optional.isPresent()) {
